requirements.txt
pywin32==311
pystray==0.19.0
Pillow==10.0.0
psutil==5.9.5
